# docs/api/openapi.v2.json
{
  "components": {
    "schemas": {
...
      "AdminPeopleListResponseV2": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/AdminPersonSummaryV2"
            },
            "title": "People",
            "type": "array"
          },
          "total_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "required": [
          "people",
          "limit",
          "offset",
          "count",
          "has_more"
        ],
        "title": "AdminPeopleListResponseV2",
        "type": "object"
      },
      "AdminPeopleReadProblemDetailV2": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "message",
          "trace_id",
          "request_id"
        ],
        "title": "AdminPeopleReadProblemDetailV2",
        "type": "object"
      },
      "AdminPeopleReadProblemResponseV2": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/AdminPeopleReadProblemDetailV2"
          }
        },
        "required": [
          "detail"
        ],
        "title": "AdminPeopleReadProblemResponseV2",
        "type": "object"
      },
...
      "AdminPersonSummaryV2": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "external_ids": {
            "additionalProperties": true,
            "title": "External Ids",
            "type": "object"
          },
          "full_name": {
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "known_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Known For"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "created_at",
          "updated_at"
        ],
        "title": "AdminPersonSummaryV2",
        "type": "object"
      },
...
      "PersonCoverPhotoDeleteResponseV2": {
        "additionalProperties": false,
        "properties": {
          "removed": {
            "title": "Removed",
            "type": "boolean"
          },
          "success": {
            "const": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "removed"
        ],
        "title": "PersonCoverPhotoDeleteResponseV2",
        "type": "object"
      },
      "PersonCoverPhotoResponseV2": {
        "additionalProperties": false,
        "properties": {
          "coverPhoto": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonCoverPhotoV2"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "coverPhoto"
        ],
        "title": "PersonCoverPhotoResponseV2",
        "type": "object"
      },
      "PersonCoverPhotoV2": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_firebase_uid": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Created By Firebase Uid",
            "type": "string"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "photo_id": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Photo Id",
            "type": "string"
          },
          "photo_url": {
            "maxLength": 8192,
            "minLength": 1,
            "title": "Photo Url",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "person_id",
          "photo_id",
          "photo_url",
          "created_at",
          "updated_at",
          "created_by_firebase_uid"
        ],
        "title": "PersonCoverPhotoV2",
        "type": "object"
      },
...
      "PersonMediaProblemDetailV2": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "message",
          "trace_id",
          "request_id"
        ],
        "title": "PersonMediaProblemDetailV2",
        "type": "object"
      },
      "PersonMediaProblemResponseV2": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/PersonMediaProblemDetailV2"
          }
        },
        "required": [
          "detail"
        ],
        "title": "PersonMediaProblemResponseV2",
        "type": "object"
      },
...
      "PersonThumbnailCropWriteResultV2": {
        "additionalProperties": false,
        "properties": {
          "link_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Id"
          },
          "origin": {
            "enum": [
              "cast_photos",
              "media_links"
            ],
            "title": "Origin",
            "type": "string"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "photo_id": {
            "format": "uuid",
            "title": "Photo Id",
            "type": "string"
          },
          "thumbnail_crop_mode": {
            "anyOf": [
              {
                "enum": [
                  "auto",
                  "manual"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Crop Mode"
          },
          "thumbnail_focus_x": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Focus X"
          },
          "thumbnail_focus_y": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Focus Y"
          },
          "thumbnail_zoom": {
            "anyOf": [
              {
                "maximum": 4.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Zoom"
          }
        },
        "required": [
          "origin",
          "photo_id",
          "person_id",
          "link_id",
          "thumbnail_focus_x",
          "thumbnail_focus_y",
          "thumbnail_zoom",
          "thumbnail_crop_mode"
        ],
        "title": "PersonThumbnailCropWriteResultV2",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "InternalAdminBearer": {
        "bearerFormat": "JWT",
        "description": "An allowlisted Supabase admin JWT or a TRR-signed internal-admin JWT. The backend applies the same InternalAdminUser authorization policy to both.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for The Reality Report - reality TV data and surveys",
    "title": "The Reality Report API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/admin/people": {
      "get": {
        "operationId": "listAdminPeopleV2",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleListResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people read request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people read request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPeopleReadProblemResponseV2"
                }
              }
            },
            "description": "The people store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "List or prefix-search core people",
        "tags": [
          "admin-core-people-reads-v2"
        ]
      }
    },
...
    "/api/v2/admin/people/{person_id}/cover-photos": {
      "delete": {
        "operationId": "deleteAdminPersonCoverPhotoV2",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonCoverPhotoDeleteResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "Delete Person Cover Photo",
        "tags": [
          "admin-person-media-v2"
        ]
      },
      "get": {
        "operationId": "getAdminPersonCoverPhotoV2",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonCoverPhotoResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "Get Person Cover Photo",
        "tags": [
          "admin-person-media-v2"
        ]
      },
      "put": {
        "operationId": "putAdminPersonCoverPhotoV2",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "photo_id": {
                    "maxLength": 2048,
                    "minLength": 1,
                    "type": "string"
                  },
                  "photo_url": {
                    "format": "uri",
                    "maxLength": 8192,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "photo_id",
                  "photo_url"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonCoverPhotoResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request is invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "Put Person Cover Photo",
        "tags": [
          "admin-person-media-v2"
        ]
      }
    },
...
    "/api/v2/admin/people/{person_id}/thumbnail-crops": {
      "put": {
        "operationId": "putAdminPersonThumbnailCropV2",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "crop": {
                    "anyOf": [
                      {
                        "additionalProperties": false,
                        "properties": {
                          "mode": {
                            "enum": [
                              "manual",
                              "auto"
                            ],
                            "type": "string"
                          },
                          "x": {
                            "maximum": 100,
                            "minimum": 0,
                            "type": "number"
                          },
                          "y": {
                            "maximum": 100,
                            "minimum": 0,
                            "type": "number"
                          },
                          "zoom": {
                            "maximum": 4,
                            "minimum": 1,
                            "type": "number"
                          }
                        },
                        "required": [
                          "x",
                          "y",
                          "zoom",
                          "mode"
                        ],
                        "type": "object"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "link_id": {
                    "anyOf": [
                      {
                        "format": "uuid",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "origin": {
                    "enum": [
                      "cast_photos",
                      "media_links"
                    ],
                    "type": "string"
                  },
                  "photo_id": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "origin",
                  "photo_id",
                  "link_id",
                  "crop"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonThumbnailCropWriteResultV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request is invalid."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The owned person photo was not found."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media request failed."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonMediaProblemResponseV2"
                }
              }
            },
            "description": "The person-media store is unavailable."
          }
        },
        "security": [
          {
            "InternalAdminBearer": []
          }
        ],
        "summary": "Put Person Thumbnail Crop",
        "tags": [
          "admin-person-media-v2"
        ]
      }
    },
...
  }
}